+++ src/lib/modules/poe.ninja/currencyView/currencySubModules/submodules/CurrencyModule.ts
@@ -1,4 +1,4 @@
-import PoeNinja from "../../../../AbstractClass/PoeNinja";
+import PoeNinja from "../../../../../AbstractClass/PoeNinja";
 import getQuickCurrency from "../../../func/getQuickCurrency";
 
 /**
@@ -10,7 +10,10 @@ class CurrencyModule extends PoeNinja {
    * @param league - The name of the league.
    * @param typeName - The name of the type.
    */
-  constructor(protected league: string, protected typeName: string) {
+  constructor(
+    protected league: string,
+    protected typeName: string,
+  ) {
     const type: string = "Currency";
     super(league, typeName, type);
   }
@@ -21,11 +24,20 @@ class CurrencyModule extends PoeNinja {
    * @returns A promise that resolves to an object containing the currency type name and its chaos equivalent value.
    * @throws If there is an error fetching the quick currency data.
    */
-  async getQuickCurrency(currencyTypeName: string = "Divine Orb"): Promise<{ currencyTypeName: string, chaosEquivalent: number }> {
+  async getQuickCurrency(
+    currencyTypeName: string = "Divine Orb",
+  ): Promise<{ currencyTypeName: string; chaosEquivalent: number }> {
     try {
-      return await getQuickCurrency(this.league, this.typeName, this.type, currencyTypeName);
+      return await getQuickCurrency(
+        this.league,
+        this.typeName,
+        this.type,
+        currencyTypeName,
+      );
     } catch (error: any) {
-      throw new Error(`Error fetching QuickCurrencyData ${this.type} and Currency Name:${currencyTypeName} data: ${error.message}`);
+      throw new Error(
+        `Error fetching QuickCurrencyData ${this.type} and Currency Name:${currencyTypeName} data: ${error.message}`,
+      );
     }
   }
 }