+++ src/lib/modules/poe.watch/allModules/subModules/CurrencyModule.ts
@@ -1,10 +1,9 @@
-import PoeWatch from "../../../AbstractClass/PoeWatch";
+import PoeWatch from "../../../../AbstractClass/PoeWatch";
 
 /**
  * Subclass of PoeWatch for accessing currency data.
  */
 class CurrencyModule extends PoeWatch {
-
   /**
    * Creates a new instance of CurrencyModule.
    * @param league - The league from which the data will be fetched.