+++ src/lib/modules/poe.watch/allModules/subModules/OilModule.ts
@@ -1,4 +1,4 @@
-import PoeWatch from "../../../AbstractClass/PoeWatch";
+import PoeWatch from "../../../../AbstractClass/PoeWatch";
 
 /**
  * Subclass of PoeWatch for accessing oil data.